Refactor java package
