+++ src/main/java/me/rexk/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package me.rexk;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
