+++ src/main/java/me/rexk/IndexController.java
@@ -1,4 +1,4 @@
-package hello;
+package me.rexk;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
